feat: migrate the rest of commands, remove manager
