+++ package.json
@@ -45,13 +45,6 @@
       "disabledByDefault": true
     },
     {
-      "name": "manager",
-      "title": "Managers",
-      "description": "Find manager profiles, career achievements, and teams managed.",
-      "mode": "view",
-      "disabledByDefault": true
-    },
-    {
       "name": "award",
       "title": "Awards",
       "description": "View a list of Premier League individual and team awards, including winners and statistics.",
