+++ src/player.tsx
@@ -3,23 +3,21 @@ import { usePromise } from "@raycast/utils";
 import groupBy from "lodash.groupby";
 import { useState } from "react";
 import { getPlayersWithTerms } from "./api";
-import { PositionSection } from "./components/player";
-import { positionMap } from "./utils";
+import { getProfileImg, positions } from "./utils";
 
 export default function EPLPlayer() {
   const [terms, setTerms] = useState<string>("");
 
   const { isLoading, data, pagination } = usePromise(
-    (terms) =>
-      async ({ page = 0 }) => {
-        return terms && terms.length >= 3
-          ? await getPlayersWithTerms(terms, page)
-          : { data: [], hasMore: false };
-      },
+    (terms) => async () => {
+      return terms && terms.length >= 3
+        ? await getPlayersWithTerms(terms)
+        : { data: [], hasMore: false };
+    },
     [terms],
   );
 
-  const positions = groupBy(data, "info.position");
+  const positionMap = groupBy(data, "data.otherFields.position");
 
   return (
     <Grid
@@ -36,15 +34,26 @@ export default function EPLPlayer() {
           title="Please enter a search term with at least 3 characters."
         />
       ) : (
-        Object.entries(positionMap).map(([key, value]) => {
-          const players = positions[key] || [];
+        positions.map((position) => {
+          const players = positionMap[position] || [];
 
           return (
-            <Grid.Section
-              key={key}
-              title={value}
-              children={PositionSection(players)}
-            />
+            <Grid.Section key={position} title={position}>
+              {players.map((player) => {
+                return (
+                  <Grid.Item
+                    key={player.data.objectId}
+                    title={player.data.otherFields.knownName}
+                    subtitle={player.data.otherFields.teamName}
+                    keywords={[player.data.otherFields.knownName]}
+                    content={{
+                      source: getProfileImg(player.data.objectSid),
+                      fallback: "player-missing.png",
+                    }}
+                  />
+                );
+              })}
+            </Grid.Section>
           );
         })
       )}
