+++ src/types/index.d.ts
@@ -3,348 +3,284 @@ export interface EPLContent<T> {
   pageInfo: PageInfo;
 }
 
-export interface EPLFixtureEvents {
-  events: EPLContent<FixtureEvent>;
-  fixture: Fixture;
+export interface EPLPagination<T> {
+  pagination: Pagination;
+  data: T[];
 }
 
-export interface FixtureEvent {
-  assistId: number;
-  clock?: Timestamp;
-  description: string;
-  id: number;
-  persionId: number;
-  phase?: string;
-  playerIds?: number[];
-  score: Score;
-  teamId: number;
-  text: string;
-  time?: Timestamp;
-  type: string;
+export interface EPLMetadata {
+  metadata: Metadata;
 }
 
 export interface EPLAward {
-  compSeason: CompSeason;
-  monthAwards: Record<string, Award[]>;
-  seasonAwards: Award[];
+  manager_awards: Player[];
+  player_awards: Player[];
 }
 
-export interface EPLStaff {
-  compSeason: CompSeason;
-  officials: Player[];
+export interface EPLPlayerStats {
+  stats: { [key: string]: number };
+  player: Player;
+}
+
+export interface EPLStandings {
+  matchweek: number;
+  tables: Table[];
+  season: SeasonInfo;
+  competition: Competition;
+  deductions: unknown[];
+  live: boolean;
+}
+
+export interface EPLMatchEvents {
+  awayTeam: TeamEvents;
+  homeTeam: TeamEvents;
+}
+
+export interface EPLMatchLineups {
+  home_team: TeamLineup;
+  away_team: TeamLineup;
+}
+
+export interface EPLMatchOfficials {
+  matchId: string;
+  matchOfficials: MatchOfficial[];
+}
+
+export interface EPLTeamSquad {
   players: Player[];
+  id: EPLCompetition;
   team: Team;
 }
 
-export interface EPLPlayer {
-  entity: Player;
-  stats: Stat[];
+export interface Career {
+  seasonsInPremierLeague: string[];
+  firstPremierLeagueFixtureId: string;
+  seasonsAtCurrentTeam: string[];
 }
 
-export interface Stat {
-  additionalInfo: unknown;
-  description: string;
+export interface Country {
+  isoCode: string;
+  country: string;
+  demonym: null | string;
+}
+
+export interface Team {
   name: string;
-  value: number;
+  id: string;
+  shortName: string;
+  abbr: string;
+  loan?: number;
+  score: number;
+  halfTimeScore: number;
+  redCards: number;
 }
 
-export interface EPLStanding {
-  compSeason: CompSeason;
-  dynamicallyGenerated: boolean;
-  live: boolean;
-  tables: Table[];
-  timestamp: Timestamp;
+export interface Dates {
+  joinedClub: Date;
+  birth: Date;
 }
 
-export interface Award {
-  apiTeam?: Team;
-  award: string;
-  awardTypeId: number;
-  official?: Player;
-  player?: Player;
-  relatedContent: RelatedContent[];
-}
-
-export interface RelatedContent {
-  reference: string;
+export interface Player {
+  currentTeam: Team;
+  date: string;
+  country: Country;
+  shirtNum: string | number;
+  weight: number;
+  dates: Dates;
   type: string;
-}
-
-export interface Fixture {
-  _source: string;
-  altIds: AltIDS;
-  attendance?: number;
-  behindClosedDoors: boolean;
-  clock?: Timestamp;
-  events: FixtureEvent[];
-  extraTime: boolean;
-  fixtureType: string;
-  gameweek: Gameweek;
-  goals: FixtureEvent[];
-  ground: Ground;
-  halfTimeScore: Score;
-  id: number;
-  kickoff: Timestamp;
-  matchOfficials: MatchOfficial[];
-  metadata: Record<string, unknown>;
-  neutralGround: boolean;
-  outcome: string;
-  penaltyShootouts: unknown[];
-  phase: string;
-  provisionalKickoff: Timestamp;
-  replay: boolean;
-  shootout: boolean;
-  status: string;
-  teamLists: TeamList[];
-  teams: TeamScore[];
-}
-
-export interface TeamList {
-  formation?: Formation;
-  lineup: Player[];
-  substitutes: Player[];
-  teamId: number;
-}
-
-export interface Formation {
-  label: string;
-  players: Array<number[]>;
-}
-
-export interface MatchOfficial {
-  birth: Birth;
-  id: number;
-  matchOfficialId: number;
-  name: Name;
+  countryOfBirth: string;
+  name: AwardName;
+  id: string;
+  position: string;
+  preferredFoot?: string;
+  height: number;
+  career: Career;
   role: string;
+  firstName: string;
+  lastName: string;
+  isCaptain: boolean;
+  knownName?: string;
+  subPosition?: string;
 }
 
-export interface Score {
-  awayScore: number;
-  homeScore: number;
+export interface AwardName {
+  last: string;
+  display: string;
+  first: string;
+  known?: string;
+}
+
+export interface Competition {
+  code: string;
+  name: string;
+  id: string;
 }
 
 export interface Table {
   entries: Entry[];
-  gameWeek: number;
 }
 
 export interface Entry {
-  annotations?: Annotation[];
-  away: Stats;
-  form: Fixture[];
-  ground: Ground;
-  home: Stats;
-  next: Fixture;
-  overall: Stats;
-  position: number;
-  startingPosition: number;
+  away: TeamStats;
+  overall: TeamStats;
   team: Team;
+  home: TeamStats;
+  form?: Fixture[];
+  next?: Fixture;
+}
+
+export interface TeamStats {
+  goalsFor: number;
+  lost: number;
+  won: number;
+  position: number;
+  drawn: number;
+  goalsAgainst: number;
+  played: number;
+  points: number;
+  startingPosition?: number;
+}
+
+export interface TeamForm extends Team {
+  next: Fixture;
+  form: Fixture[];
+}
+
+export interface Fixture {
+  kickoffTimezone: string;
+  competitionId: string;
+  period: string;
+  matchWeek: number;
+  kickoff: string;
+  awayTeam: Team;
+  seasonInfo: SeasonInfo;
+  competition: string;
+  clock: string;
+  kickoffTimezoneString: string;
+  season: string;
+  homeTeam: Team;
+  ground: string;
+  resultType: string;
+  matchId: string;
+  attendance?: number;
+}
+
+export interface SeasonInfo {
+  name: string;
+  id: string;
+}
+
+export interface EPLCompetition {
+  competitionId: string;
+  seasonId: string;
+  competitionLabel: string;
+  seasons: Season[];
+}
+
+export interface Season {
+  seasonId: string;
+  label: string;
+  annotations: Annotation[];
+  qualification: Qualification[];
+  relegation: string[];
 }
 
 export interface Annotation {
-  type: string;
-  destination: string;
-  description: string;
+  teamId?: string;
+  comment: string;
 }
 
-export interface Stats {
-  drawn: number;
-  goalsAgainst: number;
-  goalsDifference: number;
-  goalsFor: number;
-  lost: number;
-  played: number;
-  points: number;
-  won: number;
-}
-
-export interface AltIDS {
-  opta: string;
-}
-
-export interface Gameweek {
-  competitionPhase: CompetitionPhase;
-  compSeason: CompSeason;
-  gameweek: number;
-  id: number;
-}
-
-export interface CompSeason {
-  competition: Competition;
-  id: number;
+export interface Qualification {
+  competitionId: string;
   label: string;
-}
-
-export interface Competition {
-  abbreviation: string;
-  altIds: AltIDS;
-  description: string;
-  id: number;
-  level: string;
-  source: string;
-}
-
-export interface CompetitionPhase {
-  gameweekRange: number[];
-  id: number;
-  type: string;
-}
-
-export interface Ground {
-  capacity?: number;
-  city: string;
-  id: number;
-  location?: Location;
-  name: string;
-  source: string;
-}
-
-export interface Location {
-  latitude: number;
-  longitude: number;
-}
-
-export interface TeamScore {
-  score: number;
-  team: Team;
-}
-
-export interface Team {
-  altIds: AltIDS;
-  club: Club;
-  grounds: Ground[];
-  id: number;
-  metadata: {
-    club_highlights_internal_description: string;
-    club_highlights_internal_url: string;
-    communities_facebook: string;
-    communities_instagram: string;
-    communities_twitter: string;
-    communities_URL: string;
-    communities_youtube: string;
-  };
-  name: string;
-  shortName: string;
-  teamType: string;
+  positions: string[];
+  teamIds?: string[];
 }
 
 export interface Club {
-  abbr: string;
-  id: number;
   name: string;
-  shortName: string;
-}
-
-export interface PageInfo {
-  numEntries: number;
-  numPages: number;
-  page: number;
-  pageSize: number;
-}
-
-export interface Timestamp {
-  completeness?: number;
-  gmtOffset?: number;
-  label?: string;
-  millis?: number;
-  secs?: number;
-}
-
-export interface Player {
-  active: boolean;
-  age: string;
-  altIds: AltIDS;
-  appearances: number;
-  assists?: number;
-  awards?: Record<string, PlayerAward[]>;
-  birth: Birth;
-  captain: boolean;
-  cleanSheets: number;
-  currentTeam?: Team;
-  // debut: Timestamp;
-  goals?: number;
-  goalsConceded?: number;
-  height?: number;
-  id: number;
-  info: Info;
-  joinDate?: Timestamp;
-  keyPasses?: number;
-  latestPosition: string;
-  // leaveDate: null;
-  matchPosition: Position;
-  matchShirtNumber: number;
-  // metadata: null;
-  name: Name;
-  nationalTeam?: Country;
-  officialId: number;
-  playerId: number;
-  previousTeam?: Team;
-  role: string;
-  saves?: number;
-  shots?: number;
-  tackles?: number;
-  // teamHistory: null;
-  weight?: number;
-}
-
-export interface Birth {
-  country: Country;
-  date: Timestamp;
-  place?: string;
-}
-
-export interface Country {
-  country: string;
-  demonym?: string;
-  isoCode: string;
-}
-
-export interface Info {
-  loan?: boolean;
-  position: string;
-  positionInfo: string;
-  shirtNum?: number;
-}
-
-export interface Name {
-  display: string;
-  first: string;
-  last: string;
-  middle?: string;
-}
-
-export interface PlayerAward {
-  compSeason: CompSeason;
-  date: Date | number[];
-  notes?: string;
-}
-
-export interface Date {
-  day: number;
-  month: number;
-  year: number;
-}
-
-export interface EPLPlayerSearch {
-  hits: Hits;
-  status: string;
-}
-
-export interface Hits {
-  cursor: null;
-  found: number;
-  hit: Hit[];
-  start: number;
-}
-
-export interface Hit {
-  contentType: string;
+  stadium: Stadium;
   id: string;
-  response: Player;
+  shortName: string;
+  abbr: string;
 }
 
-export interface Report {
+export interface Stadium {
+  country: string;
+  city: string;
+  name: string;
+  capacity: number;
+}
+
+export interface Pagination {
+  _limit: number;
+  _prev: null;
+  _next: null;
+}
+
+export interface TeamLineup {
+  players: string[];
+  teamId: string;
+  formation: Formation;
+  managers: string[];
+}
+
+export interface Formation {
+  subs: string[];
+  teamId: string;
+  lineup: Array<string[]>;
+  formation: string;
+}
+
+export interface TeamEvents {
+  cards: MatchEvent[];
+  subs: MatchEvent[];
+  name: string;
+  id: string;
+  shortName: string;
+  goals: MatchEvent[];
+}
+
+export interface MatchEvent {
+  period: string;
+  time: string;
+  type: string;
+  playerId: string;
+  timestamp: string;
+  goalType: string;
+  assistPlayerId: null | string;
+  playerOnId: string;
+  playerOffId: string;
+}
+
+export interface MatchCommentary {
+  player1: string;
+  team1: string;
+  comment: string;
+  time: string;
+  type: string;
+  timestamp: Date;
+  player2?: string;
+  team2?: string;
+}
+
+export interface MatchOfficial {
+  official: Official;
+  type: string;
+}
+
+export interface Official {
+  firstName: string;
+  lastName: string;
+  name: string;
+}
+
+export interface EPLContentReport {
+  pageInfo: PageInfo;
+  content: Content[];
+}
+
+export interface Content {
   id: number;
   accountId: number;
   type: string;
@@ -362,19 +298,13 @@ export interface Report {
   language: string;
   additionalInfo: unknown;
   canonicalUrl: string;
-  references: Tag[];
+  references: Reference[];
   related: unknown[];
-  metadata: {
-    awayTeamMatchReportLinkText: string;
-    awayTeamMatchReportURL: string;
-    awayTeamOptaID: string;
-    homeTeamMatchReportLinkText: string;
-    homeTeamMatchReportURL: string;
-    homeTeamOptaID: string;
-  };
+  metadata: ContentMetadata;
   titleTranslations: null;
   lastModified: number;
   titleUrlSegment: string;
+  visibilityOrder: unknown[];
   body: string;
   author: null;
   subtitle: null;
@@ -387,9 +317,79 @@ export interface Report {
   onDemandUrl: null;
 }
 
-export interface Tag {
+export interface ConentMetadata {
+  awayTeamMatchReportLinkText: string;
+  awayTeamMatchReportURL: string;
+  awayTeamOptaID: string;
+  homeTeamMatchReportLinkText: string;
+  homeTeamMatchReportURL: string;
+  homeTeamOptaID: string;
+}
+
+export interface Reference {
   label: null;
   id: number;
   type: string;
   sid: string;
+}
+
+export interface Tag {
+  id: number;
+  label: string;
+}
+
+export interface PageInfo {
+  page: number;
+  numPages: number;
+  pageSize: number;
+  numEntries: number;
+}
+
+export interface EPLPlayerSearch {
+  found: number;
+  hits: Hit[];
+}
+
+export interface Hit {
+  contentType: string;
+  data: Data;
+  contentReference: null;
+}
+
+export interface Data {
+  context: string;
+  type: string;
+  language: string;
+  objectId: number;
+  title: string;
+  tags: unknown[];
+  customFields: unknown;
+  visibility: string;
+  objectSid: string;
+  otherFields: OtherFields;
+  imageUrl: unknown[];
+}
+
+export interface OtherFields {
+  teamName: string;
+  country: string;
+  user_group_restricted: string;
+  teamId: string;
+  last_name: string;
+  teamShortName: string;
+  position: string;
+  teamAbbr: string;
+  first_name: string;
+  knownName: string;
+}
+
+export interface Metadata {
+  club_established: string;
+  club_instagram_handle: string;
+  club_shop: string;
+  club_stadium: string;
+  club_tiktok_handle: string;
+  club_website: string;
+  club_x_handle: string;
+  shirt_url: string;
 }
